refactor(main): extract shard s2 union (aux_config_cmd + model_picker) from main.py
