+++ hermes_cli/main.py
@@ -802,6 +802,29 @@ from hermes_cli.model_setup_flows import (
     _model_flow_moa,
     _model_flow_ai_gateway,
 )
+
+# Provider picker / auxiliary-config flows extracted to
+# hermes_cli/model_picker.py and hermes_cli/aux_config_cmd.py (god-file
+# decomposition, wave 1 shard s2). Re-imported here so existing call sites
+# and test monkeypatches (hermes_cli.main.*) keep resolving unchanged.
+from hermes_cli.aux_config_cmd import (
+    _AUX_TASKS,
+    _all_aux_tasks,
+    _aux_config_menu,
+    _aux_flow_custom_endpoint,
+    _aux_flow_provider_model,
+    _aux_select_for_task,
+    _format_aux_current,
+    _reset_aux_to_auto,
+    _save_aux_choice,
+)
+from hermes_cli.model_picker import (
+    _is_profile_api_key_provider,
+    cmd_model,
+    select_provider_and_model,
+)
+
+
 logger = logging.getLogger(__name__)
 
 
@@ -3017,485 +3040,6 @@ def cmd_setup(args):
     run_setup_wizard(args)
 
 
-def cmd_model(args):
-    """Select default model — starts with provider selection, then model picker."""
-    _require_tty("model")
-    if getattr(args, "refresh", False):
-        try:
-            from hermes_cli.models import clear_provider_models_cache
-            clear_provider_models_cache()
-            print("  Cleared model picker cache.")
-        except Exception:
-            pass
-    select_provider_and_model(args=args)
-
-
-def _is_profile_api_key_provider(provider_id: str) -> bool:
-    """Return True when provider_id maps to a profile with auth_type='api_key'.
-
-    Used as a catch-all in select_provider_and_model() so that new providers
-    declared in plugins/model-providers/<name>/ automatically dispatch to _model_flow_api_key_provider
-    without requiring an explicit elif branch here.
-    """
-    try:
-        from providers import get_provider_profile
-        _p = get_provider_profile(provider_id)
-        return _p is not None and _p.auth_type == "api_key"
-    except Exception:
-        return False
-
-
-def select_provider_and_model(args=None):
-    """Core provider selection + model picking logic.
-
-    Shared by ``cmd_model`` (``hermes model``) and the setup wizard
-    (``setup_model_provider`` in setup.py).  Handles the full flow:
-    provider picker, credential prompting, model selection, and config
-    persistence.
-    """
-    from hermes_cli.auth import (
-        resolve_provider,
-        AuthError,
-        format_auth_error,
-    )
-    from hermes_cli.config import (
-        get_compatible_custom_providers,
-        load_config,
-        get_env_value,
-    )
-    from hermes_cli.providers import (
-        custom_provider_aliases,
-        custom_provider_slug,
-        resolve_provider_full,
-    )
-
-    config = load_config()
-    current_model = config.get("model")
-    if isinstance(current_model, dict):
-        current_model = current_model.get("default", "")
-    current_model = current_model or "(not set)"
-
-    # Read effective provider the same way the CLI does at startup:
-    # config.yaml model.provider > env var > auto-detect
-    config_provider = None
-    model_cfg = config.get("model")
-    if isinstance(model_cfg, dict):
-        config_provider = model_cfg.get("provider")
-
-    effective_provider = (
-        config_provider or os.getenv("HERMES_INFERENCE_PROVIDER") or "auto"
-    )
-    compatible_custom_providers = get_compatible_custom_providers(config)
-    def _named_custom_provider_map(cfg) -> dict[str, dict[str, str]]:
-        from hermes_cli.config import read_raw_config
-
-        # Build lookups of raw (un-expanded) templates keyed by a
-        # stable identity. We intentionally bypass
-        # ``get_compatible_custom_providers(read_raw_config())`` here because
-        # its ``_normalize_custom_provider_entry`` step calls ``urlparse()``
-        # on ``base_url`` and drops any entry whose ``base_url`` is itself an
-        # env-ref template (e.g. ``${NEURALWATT_API_BASE}``). Dropping those
-        # entries is exactly how env-ref preservation fails for the user
-        # config that motivated this fix.
-        raw_api_key_refs: dict[tuple, str] = {}
-        raw_base_url_refs: dict[tuple, str] = {}
-        raw_cfg = read_raw_config()
-
-        def _record_raw(
-            name: str,
-            provider_key: str,
-            model: str,
-            api_key: str,
-            base_url: str,
-        ) -> None:
-            template = str(api_key or "").strip()
-            base_template = str(base_url or "").strip()
-            name = str(name or "").strip()
-            provider_key = str(provider_key or "").strip()
-            model = str(model or "").strip()
-            # Index by every plausible identity the loaded (expanded) config
-            # might present: (name), (name, model), (provider_key), and
-            # (provider_key, model). Case-insensitive on name/provider_key so
-            # the loaded entry matches regardless of display casing.
-            identities = []
-            if name:
-                identities.extend(((name.lower(),), (name.lower(), model)))
-            if provider_key:
-                identities.extend(
-                    ((provider_key.lower(),), (provider_key.lower(), model))
-                )
-            if "${" in template:
-                for identity in identities:
-                    raw_api_key_refs.setdefault(identity, template)
-            if "${" in base_template:
-                for identity in identities:
-                    raw_base_url_refs.setdefault(identity, base_template)
-
-        raw_list = raw_cfg.get("custom_providers")
-        if isinstance(raw_list, list):
-            for raw_entry in raw_list:
-                if not isinstance(raw_entry, dict):
-                    continue
-                _record_raw(
-                    raw_entry.get("name", ""),
-                    "",
-                    raw_entry.get("model", "") or raw_entry.get("default_model", ""),
-                    raw_entry.get("api_key", ""),
-                    raw_entry.get("base_url", "")
-                    or raw_entry.get("url", "")
-                    or raw_entry.get("api", ""),
-                )
-        raw_providers = raw_cfg.get("providers")
-        if isinstance(raw_providers, dict):
-            for raw_key, raw_entry in raw_providers.items():
-                if not isinstance(raw_entry, dict):
-                    continue
-                _record_raw(
-                    raw_entry.get("name", "") or raw_key,
-                    raw_key,
-                    raw_entry.get("model", "") or raw_entry.get("default_model", ""),
-                    raw_entry.get("api_key", ""),
-                    raw_entry.get("base_url", "")
-                    or raw_entry.get("url", "")
-                    or raw_entry.get("api", ""),
-                )
-
-        def _lookup_ref(
-            refs: dict[tuple, str],
-            name: str,
-            provider_key: str,
-            model: str,
-        ) -> str:
-            name_lc = str(name or "").strip().lower()
-            pkey_lc = str(provider_key or "").strip().lower()
-            model = str(model or "").strip()
-            for identity in (
-                (pkey_lc, model),
-                (pkey_lc,),
-                (name_lc, model),
-                (name_lc,),
-            ):
-                if identity[0] and identity in refs:
-                    return refs[identity]
-            return ""
-
-        custom_provider_map = {}
-        for entry in get_compatible_custom_providers(cfg):
-            if not isinstance(entry, dict):
-                continue
-            name = (entry.get("name") or "").strip()
-            base_url = (entry.get("base_url") or "").strip()
-            if not name or not base_url:
-                continue
-            provider_key = (entry.get("provider_key") or "").strip()
-            key = custom_provider_slug(name, provider_key)
-            custom_provider_map[key] = {
-                "name": name,
-                "base_url": base_url,
-                "api_key": entry.get("api_key", ""),
-                "key_env": entry.get("key_env", ""),
-                "model": entry.get("model", ""),
-                "models": entry.get("models", {}),
-                "discover_models": entry.get("discover_models", True),
-                "api_mode": entry.get("api_mode", ""),
-                "provider_key": provider_key,
-                "api_key_ref": _lookup_ref(
-                    raw_api_key_refs, name, provider_key, entry.get("model", "")
-                ),
-                "base_url_ref": _lookup_ref(
-                    raw_base_url_refs, name, provider_key, entry.get("model", "")
-                ),
-            }
-        return custom_provider_map
-
-    def _norm_base_url(url: str) -> str:
-        return str(url or "").strip().rstrip("/").lower()
-
-    # Add user-defined custom providers from config.yaml
-    _custom_provider_map = _named_custom_provider_map(
-        config
-    )  # key → {name, base_url, api_key}
-
-    def _canonical_named_custom_key(provider_id: str) -> str:
-        requested = str(provider_id or "").strip().lower()
-        for key, provider_info in _custom_provider_map.items():
-            if requested in custom_provider_aliases(
-                provider_info.get("name", ""),
-                provider_info.get("provider_key", ""),
-            ):
-                return key
-        return provider_id
-
-    def _active_custom_key_from_base_url() -> str:
-        if effective_provider != "custom" or not isinstance(model_cfg, dict):
-            return ""
-        current_base = _norm_base_url(model_cfg.get("base_url", ""))
-        if not current_base:
-            return ""
-        for key, provider_info in _custom_provider_map.items():
-            if _norm_base_url(provider_info.get("base_url", "")) == current_base:
-                return key
-        return ""
-
-    active = _active_custom_key_from_base_url()
-    if active is None:
-        active = ""
-    if not active and effective_provider != "auto":
-        active_def = resolve_provider_full(
-            effective_provider,
-            config.get("providers"),
-            compatible_custom_providers,
-        )
-        if active_def is not None:
-            active = active_def.id
-            if active_def.source == "user-config":
-                active = _canonical_named_custom_key(active)
-        else:
-            warning = (
-                f"Unknown provider '{effective_provider}'. Check 'hermes model' for "
-                "available providers, or run 'hermes doctor' to diagnose config "
-                "issues."
-            )
-            print(f"Warning: {warning} Falling back to auto provider detection.")
-    if not active:
-        try:
-            active = resolve_provider("auto")
-        except AuthError as exc:
-            if effective_provider == "auto":
-                warning = format_auth_error(exc)
-                print(f"Warning: {warning} Falling back to auto provider detection.")
-            active = None  # no provider yet; default to first in list
-
-    # Detect custom endpoint
-    if active == "openrouter" and get_env_value("OPENAI_BASE_URL"):
-        active = "custom"
-
-    from hermes_cli.models import (
-        CANONICAL_PROVIDERS,
-        _PROVIDER_LABELS,
-        _PROVIDER_ALIASES,
-        group_providers,
-        provider_group_for_slug,
-    )
-
-    provider_labels = dict(_PROVIDER_LABELS)  # derive from canonical list
-    if active and active in _custom_provider_map:
-        active_label = _custom_provider_map[active]["name"]
-    else:
-        active_label = provider_labels.get(active, active) if active else "none"
-
-    print()
-    print(f"  Current model:    {current_model}")
-    print(f"  Active provider:  {active_label}")
-    print()
-
-    # Step 1: Provider selection.
-    #
-    # Canonical providers are folded into top-level groups (display only — see
-    # PROVIDER_GROUPS in hermes_cli/models.py). A multi-member group shows one
-    # row ("Kimi / Moonshot ▸"); picking it opens a member sub-picker that
-    # resolves back to a concrete slug, so the dispatch chain below is
-    # unchanged. Custom providers and the trailing actions stay flat.
-    canonical_descs = {p.slug: p.tui_desc for p in CANONICAL_PROVIDERS}
-    # Honor ``model_catalog.excluded_providers`` so the CLI ``hermes model``
-    # picker hides the same providers the gateway/TUI pickers do. A canonical
-    # provider is hidden if its slug OR any of its aliases appears in the
-    # exclusion list (case-insensitive), matching list_authenticated_providers'
-    # matching against hermes_id / alias / canonical slug.
-    _cli_excluded = {
-        str(p).strip().lower()
-        for p in (config.get("model_catalog", {}) or {}).get("excluded_providers") or []
-        if p
-    }
-    if _cli_excluded:
-        _alias_to_canon = _PROVIDER_ALIASES
-        _names_for: dict[str, set[str]] = {}
-        for _p in CANONICAL_PROVIDERS:
-            _names_for[_p.slug] = {_p.slug.lower()}
-        for _alias, _canon in _alias_to_canon.items():
-            _names_for.setdefault(_canon, {_canon.lower()}).add(_alias.lower())
-        _visible_slugs = [
-            p.slug for p in CANONICAL_PROVIDERS
-            if not _names_for.get(p.slug, {p.slug.lower()}) & _cli_excluded
-        ]
-    else:
-        _visible_slugs = [p.slug for p in CANONICAL_PROVIDERS]
-    grouped_rows = group_providers(_visible_slugs)
-
-    # The group/slug that should be pre-selected: the active provider's group
-    # if it's grouped, otherwise the active slug itself.
-    active_group = provider_group_for_slug(active) if active else ""
-
-    # ordered entries: (key, label, members)
-    #   members == [] → leaf row, key is a provider slug / action
-    #   members != [] → group row, key is "group:<gid>"
-    ordered: list[tuple[str, str, list[str]]] = []
-    default_idx = 0
-    for row in grouped_rows:
-        if row["kind"] == "group":
-            gid = row["group_id"]
-            group_desc = row.get("description", "")
-            label = f"{row['label']} ▸ ({group_desc})" if group_desc else f"{row['label']} ▸"
-            key = f"group:{gid}"
-            is_active = bool(active_group) and gid == active_group
-            members = row["members"]
-        else:
-            slug = row["slug"]
-            label = canonical_descs.get(slug, provider_labels.get(slug, slug))
-            key = slug
-            is_active = bool(active) and slug == active
-            members = []
-        if is_active:
-            ordered.append((key, f"{label}  ← currently active", members))
-            default_idx = len(ordered) - 1
-        else:
-            ordered.append((key, label, members))
-
-    for key, provider_info in _custom_provider_map.items():
-        name = provider_info["name"]
-        base_url = provider_info["base_url"]
-        short_url = base_url.replace("https://", "").replace("http://", "").rstrip("/")
-        saved_model = provider_info.get("model", "")
-        model_hint = f" — {saved_model}" if saved_model else ""
-        label = f"{name} ({short_url}){model_hint}"
-        if active and key == active:
-            ordered.append((key, f"{label}  ← currently active", []))
-            default_idx = len(ordered) - 1
-        else:
-            ordered.append((key, label, []))
-
-    ordered.append(("custom", "Custom endpoint (enter URL manually)", []))
-    _has_saved_custom_list = isinstance(config.get("custom_providers"), list) and bool(
-        config.get("custom_providers")
-    )
-    if _has_saved_custom_list:
-        ordered.append(("remove-custom", "Remove a saved custom provider", []))
-    ordered.append(("aux-config", "Configure auxiliary models...", []))
-    ordered.append(("cancel", "Leave unchanged", []))
-
-    provider_idx = _prompt_provider_choice(
-        [label for _, label, _ in ordered],
-        default=default_idx,
-    )
-    if provider_idx is None or ordered[provider_idx][0] == "cancel":
-        print("No change.")
-        return
-
-    selected_key = ordered[provider_idx][0]
-    selected_members = ordered[provider_idx][2]
-
-    # Group row → drill into a member sub-picker. Default to the active member
-    # if the active provider lives in this group. The descriptive text lives on
-    # the group row itself, so member rows show only their short label here.
-    if selected_members:
-        member_default = 0
-        if active in selected_members:
-            member_default = selected_members.index(active)
-        member_labels = [
-            provider_labels.get(m, m) for m in selected_members
-        ]
-        group_label = ordered[provider_idx][1].split(" ▸", 1)[0]
-        member_idx = _prompt_provider_choice(
-            member_labels,
-            default=member_default,
-            title=f"Select {group_label} provider:",
-        )
-        if member_idx is None:
-            print("No change.")
-            return
-        selected_provider = selected_members[member_idx]
-    else:
-        selected_provider = selected_key
-
-    if selected_provider == "aux-config":
-        _aux_config_menu()
-        return
-
-    # Step 2: Provider-specific setup + model selection
-    if selected_provider == "openrouter":
-        _model_flow_openrouter(config, current_model)
-    elif selected_provider == "moa":
-        _model_flow_moa(config, current_model)
-    elif selected_provider == "ai-gateway":
-        _model_flow_ai_gateway(config, current_model)
-    elif selected_provider == "nous":
-        _model_flow_nous(config, current_model, args=args)
-    elif selected_provider == "openai-codex":
-        _model_flow_openai_codex(config, current_model)
-    elif selected_provider == "xai-oauth":
-        _model_flow_xai_oauth(config, current_model, args=args)
-    elif selected_provider == "qwen-oauth":
-        _model_flow_qwen_oauth(config, current_model)
-    elif selected_provider == "minimax-oauth":
-        _model_flow_minimax_oauth(config, current_model, args=args)
-    elif selected_provider == "copilot-acp":
-        _model_flow_copilot_acp(config, current_model)
-    elif selected_provider == "copilot":
-        _model_flow_copilot(config, current_model)
-    elif selected_provider == "custom":
-        _model_flow_custom(config)
-    elif (
-        selected_provider.startswith("custom:")
-        or selected_provider in _custom_provider_map
-    ):
-        provider_info = _named_custom_provider_map(load_config()).get(selected_provider)
-        if provider_info is None:
-            print(
-                "Warning: the selected saved custom provider is no longer available. "
-                "It may have been removed from config.yaml. No change."
-            )
-            return
-        _model_flow_named_custom(config, provider_info)
-    elif selected_provider == "remove-custom":
-        _remove_custom_provider(config)
-    elif selected_provider == "anthropic":
-        _model_flow_anthropic(config, current_model)
-    elif selected_provider == "kimi-coding":
-        _model_flow_kimi(config, current_model)
-    elif selected_provider == "stepfun":
-        _model_flow_stepfun(config, current_model)
-    elif selected_provider == "bedrock":
-        _model_flow_bedrock(config, current_model)
-    elif selected_provider == "vertex":
-        _model_flow_vertex(config, current_model)
-    elif selected_provider == "azure-foundry":
-        _model_flow_azure_foundry(config, current_model)
-    elif selected_provider in {
-        "openai-api",
-        "gemini",
-        "deepseek",
-        "xai",
-        "zai",
-        "kimi-coding-cn",
-        "minimax",
-        "minimax-cn",
-        "kilocode",
-        "opencode-zen",
-        "opencode-go",
-        "alibaba",
-        "huggingface",
-        "xiaomi",
-        "arcee",
-        "gmi",
-        "nvidia",
-        "ollama-cloud",
-        "tencent-tokenhub",
-        "lmstudio",
-    } or _is_profile_api_key_provider(selected_provider):
-        _model_flow_api_key_provider(config, selected_provider, current_model)
-
-    # ── Post-switch cleanup: clear stale OPENAI_BASE_URL ──────────────
-    # When the user switches to a named provider (anything except "custom"),
-    # a leftover OPENAI_BASE_URL in ~/.hermes/.env can poison auxiliary
-    # clients that use provider:auto. Clear it proactively.  (#5161)
-    if selected_provider not in {
-        "custom",
-        "cancel",
-        "remove-custom",
-    } and not selected_provider.startswith("custom:"):
-        _clear_stale_openai_base_url()
-
-
 def _clear_stale_openai_base_url():
     """Remove OPENAI_BASE_URL from ~/.hermes/.env if the active provider is not 'custom'.
 
@@ -3525,386 +3069,6 @@ def _clear_stale_openai_base_url():
             else f"Cleared stale OPENAI_BASE_URL from .env (was: {stale_url})"
         )
 
-
-# ─────────────────────────────────────────────────────────────────────────────
-# Auxiliary model configuration
-#
-# Hermes uses lightweight "auxiliary" models for side tasks (vision analysis,
-# context compression, web extraction, session search, etc.). Each task has
-# its own provider+model pair in config.yaml under `auxiliary.<task>`.
-#
-# The UI lives behind "Configure auxiliary models..." at the bottom of the
-# `hermes model` provider picker. It does NOT re-run credential setup — it
-# only routes already-authenticated providers to specific aux tasks. Users
-# configure new providers through the normal `hermes model` flow first.
-# ─────────────────────────────────────────────────────────────────────────────
-
-# (task_key, display_name, short_description)
-_AUX_TASKS: list[tuple[str, str, str]] = [
-    ("vision", "Vision", "image/screenshot analysis"),
-    ("compression", "Compression", "context summarization"),
-    ("web_extract", "Web extract", "web page summarization"),
-    ("approval", "Approval", "smart command approval"),
-    ("mcp", "MCP", "MCP tool reasoning"),
-    ("title_generation", "Title generation", "session titles"),
-    ("memory_query_rewrite", "Memory query rewrite", "memory retrieval queries"),
-    ("tts_audio_tags", "TTS audio tags", "Gemini TTS tag insertion"),
-    ("skills_hub", "Skills hub", "skills search/install"),
-    ("triage_specifier", "Triage specifier", "kanban spec fleshing"),
-    ("kanban_decomposer", "Kanban decomposer", "task decomposition"),
-    ("profile_describer", "Profile describer", "auto profile descriptions"),
-    ("curator", "Curator", "skill-usage review pass"),
-]
-
-
-def _all_aux_tasks() -> list[tuple[str, str, str]]:
-    """Return built-in + plugin-registered auxiliary tasks for picker/menu use.
-
-    Built-in tasks come first (preserving order), followed by plugin tasks
-    sorted by key. Used by ``_aux_config_menu``, ``_reset_aux_to_auto``, and
-    display-name lookups so plugin-registered tasks (registered via
-    :meth:`hermes_cli.plugins.PluginContext.register_auxiliary_task`) appear
-    in the same surfaces as built-in ones without core knowing about them.
-    """
-    tasks = list(_AUX_TASKS)
-    try:
-        from hermes_cli.plugins import get_plugin_auxiliary_tasks
-        for entry in get_plugin_auxiliary_tasks():
-            tasks.append((entry["key"], entry["display_name"], entry["description"]))
-    except Exception:
-        # Plugin discovery failure must not break the aux config UI.
-        # Built-in tasks remain available.
-        pass
-    return tasks
-
-
-def _format_aux_current(task_cfg: dict) -> str:
-    """Render the current aux config for display in the task menu."""
-    if not isinstance(task_cfg, dict):
-        return "auto"
-    base_url = str(task_cfg.get("base_url") or "").strip()
-    provider = str(task_cfg.get("provider") or "auto").strip() or "auto"
-    model = str(task_cfg.get("model") or "").strip()
-    if base_url:
-        short = base_url.replace("https://", "").replace("http://", "").rstrip("/")
-        return f"custom ({short})" + (f" · {model}" if model else "")
-    if provider == "auto":
-        return "auto" + (f" · {model}" if model else "")
-    if model:
-        return f"{provider} · {model}"
-    return provider
-
-
-def _save_aux_choice(
-    task: str,
-    *,
-    provider: str,
-    model: str = "",
-    base_url: str = "",
-    api_key: str = "",
-) -> None:
-    """Persist an auxiliary task's provider/model to config.yaml.
-
-    Only writes the four routing fields — timeout, download_timeout, and any
-    other task-specific settings are preserved untouched. The main model
-    config (``model.default``/``model.provider``) is never modified.
-    """
-    from hermes_cli.config import load_config, save_config
-
-    cfg = load_config()
-    aux = cfg.setdefault("auxiliary", {})
-    if not isinstance(aux, dict):
-        aux = {}
-        cfg["auxiliary"] = aux
-    entry = aux.setdefault(task, {})
-    if not isinstance(entry, dict):
-        entry = {}
-        aux[task] = entry
-    entry["provider"] = provider
-    entry["model"] = model or ""
-    entry["base_url"] = base_url or ""
-    entry["api_key"] = api_key or ""
-    save_config(cfg)
-
-
-def _reset_aux_to_auto() -> int:
-    """Reset every known aux task back to auto/empty. Returns number reset.
-
-    Includes plugin-registered tasks (via ``_all_aux_tasks``) so a plugin
-    that contributed an auxiliary task gets reset alongside built-ins.
-    """
-    from hermes_cli.config import load_config, save_config
-
-    cfg = load_config()
-    aux = cfg.setdefault("auxiliary", {})
-    if not isinstance(aux, dict):
-        aux = {}
-        cfg["auxiliary"] = aux
-    count = 0
-    for task, _name, _desc in _all_aux_tasks():
-        entry = aux.setdefault(task, {})
-        if not isinstance(entry, dict):
-            entry = {}
-            aux[task] = entry
-        changed = False
-        if entry.get("provider") not in {None, "", "auto"}:
-            entry["provider"] = "auto"
-            changed = True
-        for field in ("model", "base_url", "api_key"):
-            if entry.get(field):
-                entry[field] = ""
-                changed = True
-        # Preserve timeout/download_timeout — those are user-tuned, not routing
-        if changed:
-            count += 1
-    save_config(cfg)
-    return count
-
-
-def _aux_config_menu() -> None:
-    """Top-level auxiliary-model picker — choose a task to configure.
-
-    Loops until the user picks "Back" so multiple tasks can be configured
-    without returning to the main provider menu.
-    """
-    from hermes_cli.config import load_config
-
-    while True:
-        cfg = load_config()
-        aux = cfg.get("auxiliary", {}) if isinstance(cfg.get("auxiliary"), dict) else {}
-
-        print()
-        print("  Auxiliary models — side-task routing")
-        print()
-        print("  Side tasks (vision, compression, web extraction, etc.) default")
-        print('  to your main chat model.  "auto" means "use my main model" —')
-        print("  Hermes only falls back to a lightweight backend (OpenRouter,")
-        print("  Nous Portal) if the main model is unavailable.  Override a")
-        print("  task below if you want it pinned to a specific provider/model.")
-        print()
-
-        # Build the task menu with current settings inline
-        all_tasks = _all_aux_tasks()
-        name_col = max(len(name) for _, name, _ in all_tasks) + 2
-        desc_col = max(len(desc) for _, _, desc in all_tasks) + 4
-        entries: list[tuple[str, str]] = []
-        for task_key, name, desc in all_tasks:
-            task_cfg = (
-                aux.get(task_key, {}) if isinstance(aux.get(task_key), dict) else {}
-            )
-            current = _format_aux_current(task_cfg)
-            label = (
-                f"{name.ljust(name_col)}{('(' + desc + ')').ljust(desc_col)}{current}"
-            )
-            entries.append((task_key, label))
-        entries.append(("__reset__", "Reset all to auto"))
-        entries.append(("__back__", "Back"))
-
-        idx = _prompt_provider_choice(
-            [label for _, label in entries],
-            default=0,
-        )
-        if idx is None:
-            return
-        key = entries[idx][0]
-        if key == "__back__":
-            return
-        if key == "__reset__":
-            n = _reset_aux_to_auto()
-            if n:
-                print(f"Reset {n} auxiliary task(s) to auto.")
-            else:
-                print("All auxiliary tasks were already set to auto.")
-            print()
-            continue
-        # Otherwise configure the specific task
-        _aux_select_for_task(key)
-
-
-def _aux_select_for_task(task: str) -> None:
-    """Pick a provider + model for a single auxiliary task and persist it.
-
-    Provider rows come from ``build_aux_picker_rows()`` — the shared aux-picker
-    substrate — so this surface shows exactly what every other aux picker
-    shows: authenticated built-ins, the user's own ``providers:`` /
-    ``custom_providers:`` endpoints, and providers whose credential pool is
-    temporarily exhausted. Only already-configured providers appear; users set
-    up new ones through the normal ``hermes model`` flow, then route aux tasks
-    to them here.
-    """
-    from hermes_cli.config import load_config
-    from hermes_cli.inventory import build_aux_picker_rows, format_aux_picker_entries
-
-    cfg = load_config()
-    aux = cfg.get("auxiliary", {}) if isinstance(cfg.get("auxiliary"), dict) else {}
-    task_cfg = aux.get(task, {}) if isinstance(aux.get(task), dict) else {}
-    current_provider = str(task_cfg.get("provider") or "auto").strip() or "auto"
-    current_model = str(task_cfg.get("model") or "").strip()
-    current_base_url = str(task_cfg.get("base_url") or "").strip()
-
-    display_name = next((name for key, name, _ in _all_aux_tasks() if key == task), task)
-
-    # Gather authenticated providers (has credentials + curated model list)
-    try:
-        providers = build_aux_picker_rows(
-            current_provider=current_provider,
-            current_model=current_model,
-            current_base_url=current_base_url,
-        )
-    except Exception as exc:
-        print(f"Could not detect authenticated providers: {exc}")
-        providers = []
-
-    entries: list[tuple[str, str, list[str]]] = []  # (slug, label, models)
-    # "auto" always first
-    auto_marker = (
-        "  ← current" if current_provider == "auto" and not current_base_url else ""
-    )
-    entries.append(("__auto__", f"auto (recommended){auto_marker}", []))
-
-    entries.extend(
-        format_aux_picker_entries(
-            providers,
-            current_provider=current_provider,
-            current_base_url=current_base_url,
-        )
-    )
-
-    # Custom endpoint (raw base_url)
-    custom_marker = "  ← current" if current_base_url else ""
-    entries.append(("__custom__", f"Custom endpoint (direct URL){custom_marker}", []))
-    entries.append(("__back__", "Back", []))
-
-    print()
-    print(f"  Configure {display_name} — current: {_format_aux_current(task_cfg)}")
-    print()
-
-    idx = _prompt_provider_choice([label for _, label, _ in entries], default=0)
-    if idx is None:
-        return
-    slug, _label, models = entries[idx]
-
-    if slug == "__back__":
-        return
-
-    if slug == "__auto__":
-        _save_aux_choice(task, provider="auto", model="", base_url="", api_key="")
-        print(f"{display_name}: reset to auto.")
-        return
-
-    if slug == "__custom__":
-        _aux_flow_custom_endpoint(task, task_cfg)
-        return
-
-    # Regular provider — pick a model from its curated list
-    _aux_flow_provider_model(task, slug, models, current_model)
-
-
-def _aux_flow_provider_model(
-    task: str,
-    provider_slug: str,
-    curated_models: list,
-    current_model: str = "",
-) -> None:
-    """Prompt for a model under an already-authenticated provider, save to aux."""
-    from hermes_cli.auth import _prompt_model_selection
-    from hermes_cli.models import get_pricing_for_provider
-
-    display_name = next((name for key, name, _ in _all_aux_tasks() if key == task), task)
-
-    # Fetch live pricing for this provider (non-blocking)
-    pricing: dict = {}
-    try:
-        pricing = get_pricing_for_provider(provider_slug) or {}
-    except Exception:
-        pricing = {}
-
-    model_list = list(curated_models)
-
-    # Let the user pick a model. _prompt_model_selection supports "Enter custom
-    # model name" and cancel.  When there's no curated list (rare), fall back
-    # to a raw input prompt.
-    if not model_list:
-        print(f"No curated model list for {provider_slug}.")
-        print("Enter a model slug manually (blank = use provider default):")
-        try:
-            val = input("Model: ").strip()
-        except (KeyboardInterrupt, EOFError):
-            print()
-            return
-        selected = val or ""
-    else:
-        selected = _prompt_model_selection(
-            model_list,
-            current_model=current_model,
-            pricing=pricing,
-            confirm_provider=provider_slug,
-        )
-        if selected is None:
-            print("No change.")
-            return
-
-    _save_aux_choice(
-        task, provider=provider_slug, model=selected or "", base_url="", api_key=""
-    )
-    if selected:
-        print(f"{display_name}: {provider_slug} · {selected}")
-    else:
-        print(f"{display_name}: {provider_slug} (provider default model)")
-
-
-def _aux_flow_custom_endpoint(task: str, task_cfg: dict) -> None:
-    """Prompt for a direct OpenAI-compatible base_url + optional api_key/model."""
-    from hermes_cli.secret_prompt import masked_secret_prompt
-
-    display_name = next((name for key, name, _ in _all_aux_tasks() if key == task), task)
-    current_base_url = str(task_cfg.get("base_url") or "").strip()
-    current_model = str(task_cfg.get("model") or "").strip()
-
-    print()
-    print(f"  Custom endpoint for {display_name}")
-    print("  Provide an OpenAI-compatible base URL (e.g. http://localhost:11434/v1)")
-    print()
-    try:
-        url_prompt = (
-            f"Base URL [{current_base_url}]: " if current_base_url else "Base URL: "
-        )
-        url = input(url_prompt).strip()
-    except (KeyboardInterrupt, EOFError):
-        print()
-        return
-    url = url or current_base_url
-    if not url:
-        print("No URL provided. No change.")
-        return
-    try:
-        model_prompt = (
-            f"Model slug (optional) [{current_model}]: "
-            if current_model
-            else "Model slug (optional): "
-        )
-        model = input(model_prompt).strip()
-    except (KeyboardInterrupt, EOFError):
-        print()
-        return
-    model = model or current_model
-    try:
-        api_key = masked_secret_prompt(
-            "API key (optional, blank = use OPENAI_API_KEY): "
-        ).strip()
-    except (KeyboardInterrupt, EOFError):
-        print()
-        return
-
-    _save_aux_choice(
-        task,
-        provider="custom",
-        model=model,
-        base_url=url,
-        api_key=api_key,
-    )
-    short_url = url.replace("https://", "").replace("http://", "").rstrip("/")
-    print(f"{display_name}: custom ({short_url})" + (f" · {model}" if model else ""))
 
 
 def _prompt_provider_choice(choices, *, default=0, title="Select provider:"):
